Moved logic to help figure out if we are running in Azure to a helper class
