+++ src/DependencyManagement/ManagedDependenciesPathDetector.cs
@@ -10,11 +10,6 @@ namespace Microsoft.Azure.Functions.PowerShellWorker.DependencyManagement
 {
     internal static class ManagedDependenciesPathDetector
     {
-        // Environment variables to help figure out if we are running in Azure.
-        private const string AzureWebsiteInstanceId = "WEBSITE_INSTANCE_ID";
-        private const string ContainerName = "CONTAINER_NAME";
-        private const string LegionServiceHost = "LEGION_SERVICE_HOST";
-
         private const string HomeDriveName = "HOME";
         private const string DataFolderName = "data";
 
@@ -31,13 +26,8 @@ namespace Microsoft.Azure.Functions.PowerShellWorker.DependencyManagement
         /// </summary>
         public static string GetManagedDependenciesPath(string functionAppRootPath)
         {
-            if (IsLinuxConsumptionOnLegion())
-            {
-                throw new NotSupportedException(PowerShellWorkerStrings.ManagedDependenciesIsNotSupportedOnLegion);
-            }
-
             // If we are running in Azure App Service or Linux Consumption use the 'HOME\data' path.
-            if (IsAppService() || IsLinuxConsumption())
+            if (WorkerEnvironment.IsAppService() || WorkerEnvironment.IsLinuxConsumption())
             {
                 var homeDriveVariable = Environment.GetEnvironmentVariable(HomeDriveName);
                 if (string.IsNullOrEmpty(homeDriveVariable))
@@ -53,23 +43,6 @@ namespace Microsoft.Azure.Functions.PowerShellWorker.DependencyManagement
             string functionAppName = Path.GetFileName(functionAppRootPath);
             string appDataFolder = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData, Environment.SpecialFolderOption.DoNotVerify);
             return Path.Combine(appDataFolder, AzureFunctionsFolderName, functionAppName, ManagedDependenciesFolderName);
-        }
-
-        private static bool IsAppService()
-        {
-            return !string.IsNullOrEmpty(Environment.GetEnvironmentVariable(AzureWebsiteInstanceId));
-        }
-
-        private static bool IsLinuxConsumption()
-        {
-            return !IsAppService() && !string.IsNullOrEmpty(Environment.GetEnvironmentVariable(ContainerName));
-        }
-
-        public static bool IsLinuxConsumptionOnLegion()
-        {
-            return !IsAppService() &&
-                   !string.IsNullOrEmpty(Environment.GetEnvironmentVariable(ContainerName)) &&
-                   !string.IsNullOrEmpty(Environment.GetEnvironmentVariable(LegionServiceHost));
         }
     }
 }
